+++ test/me/pmatiello/mockfn/matchers_test.clj
@@ -10,7 +10,7 @@
       (is (false? (matchers/matches? exactly 2))))
 
     (testing "provides an informative string representation"
-      (is (= "exactly 1" (matchers/description exactly))))))
+      (is (= "｢exactly 1｣" (matchers/description exactly))))))
 
 (deftest at-least-test
   (let [at-least (matchers/at-least 2)]
@@ -20,7 +20,7 @@
       (is (false? (matchers/matches? at-least 1))))
 
     (testing "provides an informative string representation"
-      (is (= "at-least 2" (matchers/description at-least))))))
+      (is (= "｢at-least 2｣" (matchers/description at-least))))))
 
 (deftest at-most-test
   (let [at-most (matchers/at-most 2)]
@@ -30,7 +30,7 @@
       (is (false? (matchers/matches? at-most 3))))
 
     (testing "provides an informative string representation"
-      (is (= "at-most 2" (matchers/description at-most))))))
+      (is (= "｢at-most 2｣" (matchers/description at-most))))))
 
 (deftest any-test
   (let [any (matchers/any)]
@@ -40,7 +40,7 @@
       (is (true? (matchers/matches? any 01234567890M))))
 
     (testing "provides an informative string representation"
-      (is (= "any" (matchers/description any))))))
+      (is (= "｢any｣" (matchers/description any))))))
 
 (deftest a-test
   (let [a (matchers/a Keyword)]
@@ -49,7 +49,7 @@
       (is (false? (matchers/matches? a "string"))))
 
     (testing "provides an informative string representation"
-      (is (= "a clojure.lang.Keyword" (matchers/description a))))))
+      (is (= "｢a clojure.lang.Keyword｣" (matchers/description a))))))
 
 (deftest str-starts-with-test
   (let [starts-with (matchers/str-starts-with "prefix")]
@@ -58,7 +58,7 @@
       (is (false? (matchers/matches? starts-with "no-prefix"))))
 
     (testing "provides an informative string representation"
-      (is (= "str-starts-with \"prefix\"" (matchers/description starts-with))))))
+      (is (= "｢str-starts-with \"prefix\"｣" (matchers/description starts-with))))))
 
 (deftest str-ends-with-test
   (let [ends-with (matchers/str-ends-with "suffix")]
@@ -67,7 +67,7 @@
       (is (false? (matchers/matches? ends-with "suffix-no"))))
 
     (testing "provides an informative string representation"
-      (is (= "str-ends-with \"suffix\"" (matchers/description ends-with))))))
+      (is (= "｢str-ends-with \"suffix\"｣" (matchers/description ends-with))))))
 
 (deftest str-includes-test
   (let [includes (matchers/str-includes "substring")]
@@ -76,7 +76,7 @@
       (is (false? (matchers/matches? includes "no match here"))))
 
     (testing "provides an informative string representation"
-      (is (= "str-includes \"substring\"" (matchers/description includes))))))
+      (is (= "｢str-includes \"substring\"｣" (matchers/description includes))))))
 
 (deftest str-rexp-test
   (let [rexp (matchers/str-rexp #"^prefix.*suffix$")]
@@ -85,4 +85,4 @@
       (is (false? (matchers/matches? rexp "no-match"))))
 
     (testing "provides an informative string representation"
-      (is (= "str-rexp #\"^prefix.*suffix$\"" (matchers/description rexp))))))
+      (is (= "｢str-rexp #\"^prefix.*suffix$\"｣" (matchers/description rexp))))))