+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project are documented in this file.
 
 ## Releases
 
-## 0.3.0 - ????-??-??
+## 0.3.0 - 2025-02-21
 
 - Improved documentation formatting.
 - Rename mockfn.macros to mockfn.plain.